+++ SpeechService/SpeechService.cs
@@ -61,14 +61,14 @@ namespace EddiSpeechService
             configuration = SpeechServiceConfiguration.FromFile();
         }
 
-        public void Say(Ship ship, string speech, bool wait, int priority = 3)
+        public void Say(Ship ship, string speech, bool wait, int priority = 3, string voice = null)
         {
             if (speech == null)
             {
                 return;
             }
 
-            Speak(speech, null, echoDelayForShip(ship), distortionLevelForShip(ship), chorusLevelForShip(ship), reverbLevelForShip(ship), 0, wait, priority);
+            Speak(speech, voice, echoDelayForShip(ship), distortionLevelForShip(ship), chorusLevelForShip(ship), reverbLevelForShip(ship), 0, wait, priority);
         }
 
         public void ShutUp()
@@ -243,15 +243,13 @@ namespace EddiSpeechService
                 Logging.Debug("Source is null; skipping");
                 return;
             }
-            Logging.Debug("Creating waitHandle");
+
             using (EventWaitHandle waitHandle = new EventWaitHandle(false, EventResetMode.AutoReset))
             {
                 ISoundOut soundOut = GetSoundOut();
                 try
                 {
-                    Logging.Debug("Setting up soundOut");
                     soundOut.Initialize(source);
-                    Logging.Debug("Configuring waitHandle");
                     soundOut.Stopped += (s, e) => waitHandle.Set();
 
                     TimeSpan waitTime = source.GetTime(source.Length);
@@ -321,18 +319,16 @@ namespace EddiSpeechService
 
                         Logging.Debug("Configuration is " + configuration == null ? "<null>" : JsonConvert.SerializeObject(configuration));
                         synth.Rate = configuration.Rate;
-                        Logging.Debug("Rate is " + synth.Rate);
                         synth.Volume = configuration.Volume;
-                        Logging.Debug("Volume is " + synth.Volume);
 
                         synth.SetOutputToWaveStream(stream);
-                        Logging.Debug("Output set to stream");
 
                         if (speech.Contains("<"))
                         {
                             Logging.Debug("Obtaining best guess culture");
                             string culture = bestGuessCulture(synth);
                             Logging.Debug("Best guess culture is " + culture);
+                            //speech = "<?xml version=\"1.0\" encoding=\"UTF-8\"?><speak version=\"1.0\" xmlns=\"http://www.w3.org/2001/10/synthesis\" xml:lang=\"" + bestGuessCulture(synth) + "\" onlangfailure=\"ignorelang\"><s>" + speech + "</s></speak>";
                             speech = "<?xml version=\"1.0\" encoding=\"UTF-8\"?><speak version=\"1.0\" xmlns=\"http://www.w3.org/2001/10/synthesis\" xml:lang=\"" + bestGuessCulture(synth) + "\"><s>" + speech + "</s></speak>";
                             Logging.Debug("Feeding SSML to synthesizer: " + speech);
                             synth.SpeakSsml(speech);