+++ VoiceAttackResponder/VoiceAttackPlugin.cs
@@ -15,6 +15,7 @@ using System.Text.RegularExpressions;
 using EddiSpeechResponder;
 using System.Windows;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace EddiVoiceAttackResponder
 {
@@ -39,8 +40,6 @@ namespace EddiVoiceAttackResponder
 
         public static BlockingCollection<Event> EventQueue = new BlockingCollection<Event>();
         public static Thread updaterThread = null;
-
-        private static SpeechResponder speechResponder = new SpeechResponder();
 
         public static void VA_Init1(dynamic vaProxy)
         {
@@ -68,8 +67,8 @@ namespace EddiVoiceAttackResponder
                 // Spin out a worker thread to keep the VoiceAttack events up-to-date and run event-specific commands
                 updaterThread = new Thread(() =>
                 {
-                        while (true)
-                        {
+                    while (true)
+                    {
                         try
                         {
                             Event theEvent = EventQueue.Take();
@@ -78,51 +77,57 @@ namespace EddiVoiceAttackResponder
                             // Update all standard values
                             setValues(ref vaProxy);
 
-                            // Event-specific values
-                            foreach (string key in Events.VARIABLES[theEvent.type].Keys)
-                            {
-                                // Obtain the value by name.  Actually looking for a method get_<name>
-                                System.Reflection.MethodInfo method = theEvent.GetType().GetMethod("get_" + key);
-                                if (method != null)
-                                {
-                                    Type returnType = method.ReturnType;
-                                    if (returnType.IsGenericType && returnType.GetGenericTypeDefinition() == typeof(Nullable<>))
-                                    {
-                                        returnType = Nullable.GetUnderlyingType(returnType);
-                                    }
+                            setJsonValues(ref vaProxy, "EDDI " + theEvent.type.ToLowerInvariant(), JsonConvert.DeserializeObject(JsonConvert.SerializeObject(theEvent)));
 
-                                    string varname = "EDDI " + theEvent.type.ToLowerInvariant() + " " + key;
+                            //// Event-specific values
+                            //foreach (string key in Events.VARIABLES[theEvent.type].Keys)
+                            //{
+                            //    // Obtain the value by name.  Actually looking for a method get_<name>
+                            //    System.Reflection.MethodInfo method = theEvent.GetType().GetMethod("get_" + key);
+                            //    if (method != null)
+                            //    {
+                            //        Type returnType = method.ReturnType;
+                            //        if (returnType.IsGenericType && returnType.GetGenericTypeDefinition() == typeof(Nullable<>))
+                            //        {
+                            //            returnType = Nullable.GetUnderlyingType(returnType);
+                            //        }
 
-                                    if (returnType == typeof(string))
-                                    {
-                                        vaProxy.SetText(varname, (string)method.Invoke(theEvent, null));
-                                    }
-                                    else if (returnType == typeof(int))
-                                    {
-                                        vaProxy.SetInt(varname, (int?)method.Invoke(theEvent, null));
-                                    }
-                                    else if (returnType == typeof(bool))
-                                    {
-                                        vaProxy.SetBoolean(varname, (bool?)method.Invoke(theEvent, null));
-                                    }
-                                    else if (returnType == typeof(decimal))
-                                    {
-                                        vaProxy.SetDecimal(varname, (decimal?)method.Invoke(theEvent, null));
-                                    }
-                                    else if (returnType == typeof(double))
-                                    {
-                                        vaProxy.SetDecimal(varname, (decimal?)(double?)method.Invoke(theEvent, null));
-                                    }
-                                    else if (returnType == typeof(long))
-                                    {
-                                        vaProxy.SetDecimal(varname, (decimal?)(long?)method.Invoke(theEvent, null));
-                                    }
-                                    else
-                                    {
-                                        Logging.Debug("Not handling event field type " + method.ReturnType);
-                                    }
-                                }
-                            }
+                            //        string varname = "EDDI " + theEvent.type.ToLowerInvariant() + " " + key;
+
+                            //        if (returnType == typeof(string))
+                            //        {
+                            //            vaProxy.SetText(varname, (string)method.Invoke(theEvent, null));
+                            //        }
+                            //        else if (returnType == typeof(int))
+                            //        {
+                            //            vaProxy.SetInt(varname, (int?)method.Invoke(theEvent, null));
+                            //        }
+                            //        else if (returnType == typeof(bool))
+                            //        {
+                            //            vaProxy.SetBoolean(varname, (bool?)method.Invoke(theEvent, null));
+                            //        }
+                            //        else if (returnType == typeof(decimal))
+                            //        {
+                            //            vaProxy.SetDecimal(varname, (decimal?)method.Invoke(theEvent, null));
+                            //        }
+                            //        else if (returnType == typeof(double))
+                            //        {
+                            //            vaProxy.SetDecimal(varname, (decimal?)(double?)method.Invoke(theEvent, null));
+                            //        }
+                            //        else if (returnType == typeof(long))
+                            //        {
+                            //            vaProxy.SetDecimal(varname, (decimal?)(long?)method.Invoke(theEvent, null));
+                            //        }
+                            //        else if (returnType == typeof(List<>))
+                            //        {
+                            //            Logging.Debug("Not handling list " + key);
+                            //        }
+                            //        else
+                            //        {
+                            //            Logging.Debug("Not handling event field type " + method.ReturnType);
+                            //        }
+                            //    }
+                            //}
 
                             // Fire local command if present
                             string commandName = "((EDDI " + theEvent.type.ToLowerInvariant() + "))";
@@ -156,7 +161,78 @@ namespace EddiVoiceAttackResponder
             }
         }
 
-        public static void VA_Exit1(dynamic vaProxy)
+        /// <summary>
+        /// Walk a JSON object and write out all of the possible fields
+        /// </summary>
+        private static void setJsonValues(ref dynamic vaProxy, string prefix, dynamic json)
+        {
+            foreach (JProperty child in json)
+            {
+                // We only take fully lower-cased entities, and ignore the raw key (as it's the raw journal event)
+                if ((!new Regex("^[a-z]+$").IsMatch(child.Name)) || child.Name == "raw")
+                {
+                    Logging.Debug("Ignoring key " + child.Name);
+                    continue;
+                }
+
+                string name = prefix + " " + child.Name;
+
+                if (child.Value == null)
+                {
+                    // No idea what it might have been so reset everything
+                    Logging.Warn(prefix + " " + child.Name + " is null; need to reset all values");
+                    vaProxy.SetText(name, null);
+                    vaProxy.SetInt(name, null);
+                    vaProxy.SetDecimal(name, null);
+                    vaProxy.SetBoolean(name, null);
+                    continue;
+                }
+                if (child.Value.Type == JTokenType.Boolean)
+                {
+                    Logging.Debug("Setting boolean value " + name);
+                    vaProxy.SetBoolean(name, (bool?)child.Value);
+                }
+                else if (child.Value.Type == JTokenType.String)
+                {
+                    Logging.Debug("Setting string value " + name);
+                    vaProxy.SetText(name, (string)child.Value);
+                }
+                else if (child.Value.Type == JTokenType.Float)
+                {
+                    Logging.Debug("Setting decimal value " + name);
+                    vaProxy.SetDecimal(name, (decimal?)(double?)child.Value);
+                }
+                else if (child.Value.Type == JTokenType.Integer)
+                {
+                    Logging.Debug("Setting integer value " + name);
+                    vaProxy.SetInt(name, (int?)(long?)child.Value);
+                }
+                else if (child.Value.Type == JTokenType.Date)
+                {
+                    Logging.Debug("Setting date value " + name);
+                    vaProxy.SetDate(name, (DateTime?)child.Value);
+                }
+                else if (child.Value.Type == JTokenType.Array)
+                {
+                    int i = 0;
+                    foreach (JToken arrayChild in child.Value.Children())
+                    {
+                        setJsonValues(ref vaProxy, prefix + " " + child.Name + " " + i++, arrayChild);
+                    }
+                    vaProxy.SetInt(name + " entries", i);
+                }
+                else if (child.Value.Type == JTokenType.Object)
+                {
+                    setJsonValues(ref vaProxy, prefix + " " + child.Name, child.Value);
+                }
+                else
+                {
+                    Logging.Warn(child.Value.Type + ": " + child.Name + "=" + child.Value);
+                }
+            }
+        }
+
+    public static void VA_Exit1(dynamic vaProxy)
         {
             Logging.Info("EDDI VoiceAttack plugin exiting");
             updaterThread.Abort();
@@ -200,14 +276,14 @@ namespace EddiVoiceAttackResponder
                     case "setstate":
                         InvokeSetState(ref vaProxy);
                         break;
-                    default:
-                        //if (context.ToLower().StartsWith("event:"))
-                        //{
-                        //    // Inject an event
-                        //    string data = context.Replace("event: ", "");
-                        //    JObject eventData = JObject.Parse(data);
-                        //    //LogQueue.Add(eventData);
-                        //}
+                    case "disablespeechresponder":
+                        InvokeDisableSpeechResponder(ref vaProxy);
+                        break;
+                    case "enablespeechresponder":
+                        InvokeEnableSpeechResponder(ref vaProxy);
+                        break;
+                    case "setspeechresponderpersonality":
+                        InvokeSetSpeechResponderPersonality(ref vaProxy);
                         break;
                 }
             }
@@ -419,9 +495,11 @@ namespace EddiVoiceAttackResponder
                     priority = 3;
                 }
 
+                string voice = vaProxy.GetText("Voice");
+
                 string speech = SpeechFromScript(script);
 
-                SpeechService.Instance.Say(EDDI.Instance.Ship, speech, true, (int)priority);
+                SpeechService.Instance.Say(EDDI.Instance.Ship, speech, true, (int)priority, voice);
             }
             catch (Exception e)
             {
@@ -457,11 +535,60 @@ namespace EddiVoiceAttackResponder
 
                 int? priority = vaProxy.GetInt("Priority");
 
-                speechResponder.Say(script, null, priority);
+                SpeechResponder speechResponder = (SpeechResponder)EDDI.Instance.ObtainResponder("Speech responder");
+                if (speechResponder == null)
+                {
+                    Logging.Warn("Unable to find speech responder");
+                }
+
+                string voice = vaProxy.GetText("Voice");
+
+                speechResponder.Say(script, null, priority, voice);
             }
             catch (Exception e)
             {
                 setStatus(ref vaProxy, "Failed to run internal speech system", e);
+            }
+        }
+
+        public static void InvokeDisableSpeechResponder(ref dynamic vaProxy)
+        {
+            try
+            {
+                EDDI.Instance.DisableResponder("Speech responder");
+            }
+            catch (Exception e)
+            {
+                setStatus(ref vaProxy, "Failed to disable speech responder", e);
+            }
+        }
+
+        public static void InvokeEnableSpeechResponder(ref dynamic vaProxy)
+        {
+            try
+            {
+                EDDI.Instance.EnableResponder("Speech responder");
+            }
+            catch (Exception e)
+            {
+                setStatus(ref vaProxy, "Failed to enable speech responder", e);
+            }
+        }
+
+        public static void InvokeSetSpeechResponderPersonality(ref dynamic vaProxy)
+        {
+            string personality = vaProxy.GetText("Personality");
+            try
+            {
+                SpeechResponder speechResponder = (SpeechResponder)EDDI.Instance.ObtainResponder("Speech responder");
+                if (speechResponder != null)
+                {
+                    speechResponder.SetPersonality(personality);
+                }
+            }
+            catch (Exception e)
+            {
+                setStatus(ref vaProxy, "Failed to set speech responder personality", e);
             }
         }